chnage profile page to livewirw page
