--- app/Http/Livewire/Profile/Profile.php
@@ -1,13 +1,0 @@
-<?php
-
-namespace App\Http\Livewire\Profile;
-
-use Livewire\Component;
-
-class Profile extends Component
-{
-    public function render()
-    {
-        return view('livewire.profile.profile');
-    }
-}
