--- resources/views/livewire/profile/profile.blade.php
@@ -1,3 +1,0 @@
-<div>
-    {{-- The best athlete wants his opponent at his best. --}}
-</div>
